Added alarm feature

+++ garden-service/src/garden_service/DataPoint.java
@@ -1,17 +1,17 @@
 package garden_service;
 
 class DataPoint {
-	private double temperature;
+	private int temperature;
 	private int lux;
 	private String modality;
 	
-	public DataPoint(double temperature, int lux, String modality) {
+	public DataPoint(int temperature, int lux, String modality) {
 		this.temperature = temperature;
 		this.lux = lux;
 		this.modality = modality;
 	}
 	
-	public double getTemperature() {
+	public int getTemperature() {
 		return temperature;
 	}
